Refactored the charm and moved common things in the charm base
