+++ requirements.txt
@@ -1,7 +1,8 @@
 bcrypt==4.0.1
-cryptography==38.0.3
-jsonschema==4.17.0
-ops==1.5.3
+cryptography==38.0.4
+jsonschema==4.17.3
+ops==1.5.4
+overrides==7.3.1
 requests==2.28.1
 ruamel.yaml==0.17.21
 tenacity==8.1.0
